% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{hc_set_title}
\alias{hc_set_title}
\title{Set a new title or subtitle for the chart.}
\usage{
hc_set_title(proxy, title_opts = NULL, subtitle_opts = NULL,
  redraw = FALSE)
}
\arguments{
\item{title_opts}{New title options.}

\item{subtitle_opts}{New subtitle options.}

\item{redraw}{Whether to redraw the chart or wait for a later call.}
}
\description{
Set a new title or subtitle for the chart.
}
